+++ src/compas_blender/artists/__init__.py
@@ -5,7 +5,11 @@ artists
 
 .. currentmodule:: compas_blender.artists
 
-Artists for visualizing (painting) COMPAS data structures in Blender.
+.. rst-class:: lead
+
+Artists for visualizing (painting) COMPAS geometry, robots, and data structures in Blender.
+The artists in this package provide plugins for the pluggable methods of the base artists defined in :mod:`compas.artists`.
+They can be used directly from here (:mod:`compas_blender.artists`), or through the base artist :class:`compas.artists.Artist`.
 
 
 Primitive Artists
@@ -13,6 +17,7 @@ Primitive Artists
 
 .. autosummary::
     :toctree: generated/
+    :nosignatures:
 
     CircleArtist
     FrameArtist