+++ src/compas_blender/artists/torusartist.py
@@ -13,14 +13,47 @@ from .artist import BlenderArtist
 
 
 class TorusArtist(BlenderArtist, ShapeArtist):
-    """Artist for drawing torus shapes.
+    """Artist for drawing torus shapes in Blender.
 
     Parameters
     ----------
     torus : :class:`compas.geometry.Torus`
         A COMPAS torus.
-    collection: str or :class:`bpy.types.Collection`
+    collection: str or :blender:`bpy.types.Collection`
         The name of the collection the object belongs to.
+    collection : str or :blender:`bpy.types.Collection`
+        The Blender scene collection the object(s) created by this artist belong to.
+    **kwargs : dict, optional
+        Additional keyword arguments.
+        For more info,
+        see :class:`compas_blender.artists.BlenderArtist` and :class:`compas.artists.ShapeArtist`.
+
+    Examples
+    --------
+    Use the Blender artist explicitly.
+
+    .. code-block:: python
+
+        from compas.geometry import Plane, Torus
+        from compas_blender.artists import TorusArtist
+
+        torus = Torus(Plane([0, 0, 0], [0, 0, 1]), 1.0, 0.3)
+
+        artist = TorusArtist(torus)
+        artist.draw()
+
+    Or, use the artist through the plugin mechanism.
+
+    .. code-block:: python
+
+        from compas.geometry import Plane, Torus
+        from compas.artists import Artist
+
+        torus = Torus(Plane([0, 0, 0], [0, 0, 1]), 1.0, 0.3)
+
+        artist = Artist(torus)
+        artist.draw()
+
     """
 
     def __init__(self,
@@ -35,14 +68,15 @@ class TorusArtist(BlenderArtist, ShapeArtist):
 
         Parameters
         ----------
-        color : tuple of float or tuple of int, optional
+        color : tuple[float, float, float] or tuple[int, int, int], optional
             The RGB color of the torus.
+            The default color is :attr:`compas.artists.ShapeArtist.color`.
         u : int, optional
             Number of faces in the "u" direction.
-            Default is ``~TorusArtist.u``.
+            Default is :attr:`TorusArtist.u`.
         v : int, optional
             Number of faces in the "v" direction.
-            Default is ``~TorusArtist.v``.
+            Default is :attr:`TorusArtist.v`.
 
         Returns
         -------