+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+* Removed `compas.geometry.Collection`
+* Removed `compas.geometry.CollectionNumpy`
+* Removed `compas.geometry.PointCollection`
+* Removed `compas.geometry.PointCollectionNumpy`
+
 
 ## [1.13.3] 2021-12-17